+++ earth_enterprise/src/fusion/autoingest/sysman/InNodeVertexIndexMap.h
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2019 The Open GEE Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef IN_NODE_VERTEX_INDEX_MAP_H
+#define IN_NODE_VERTEX_INDEX_MAP_H
+
+#include "DependentStateTree.h"
+
+// The depth_first_search function needs a way to map vertices to indexes. We
+// store a unique index inside each vertex; the code below provides a way for
+// boost to access them. These must be defined before including
+// depth_first_search.hpp.
+template <class Graph>
+class InNodeVertexIndexMap {
+  public:
+    typedef boost::readable_property_map_tag category;
+    typedef size_t value_type;
+    typedef value_type reference;
+    typedef typename Graph::vertex_descriptor key_type;
+
+    InNodeVertexIndexMap(const Graph & graph) : graph(graph) {};
+    const Graph & graph;
+};
+
+namespace boost {
+  template<>
+  struct property_map<DependentStateTree, vertex_index_t> {
+    typedef InNodeVertexIndexMap<DependentStateTree> const_type;
+  };
+
+  template<class Graph>
+  InNodeVertexIndexMap<Graph> get(vertex_index_t, const Graph & graph) {
+    return InNodeVertexIndexMap<Graph>(graph);
+  }
+
+  template<class Graph>
+  typename InNodeVertexIndexMap<Graph>::value_type get(
+      const InNodeVertexIndexMap<Graph> & map,
+      typename InNodeVertexIndexMap<Graph>::key_type vertex) {
+    return map.graph[vertex].index;
+  }
+}
+
+#endif
